Fix compatability issues with latest upstream merge

+++ encfs/encfs.h
@@ -73,7 +73,7 @@ int encfs_fgetattr(const char *path, struct stat_st *stbuf,
 struct fuse_file_info *fi);
 int encfs_readlink(const char *path, char *buf, size_t size);
 int encfs_readdir(const char *path, void *buf, fuse_fill_dir_t filler,
-                  off_t offset, struct fuse_file_info *finfo);
+                  long long offset, struct fuse_file_info *finfo);
 int encfs_mknod(const char *path, mode_t mode, dev_t rdev);
 int encfs_mkdir(const char *path, mode_t mode);
 int encfs_unlink(const char *path);
